--- sqs_pipeline/README.md
@@ -1,22 +1,0 @@
-# sqs_pipeline
-This is a simple application using GenStage in data processing.
-**TODO:**
-
-## Installation
-
-If [available in Hex](https://hex.pm/docs/publish), the package can be installed
-by adding `gtube` to your list of dependencies in `mix.exs`:
-
-```elixir
-def deps do
-  [
-    {:gtube, "~> 0.1.0"}
-  ]
-end
-```
-
-Documentation can be generated with [ExDoc](https://github.com/elixir-lang/ex_doc)
-and published on [HexDocs](https://hexdocs.pm). Once published, the docs can
-be found at <https://hexdocs.pm/gtube>.
-
-### The pipeline
