--- sqs_pipeline/Requirements.txt
@@ -1,13 +1,0 @@
-##Create a pipeline
-
-## S3/SQS -> [Producer] -> [Processor] -> [Consumer] -> Database
-
-## When a file is placed in S3, SQS will be notified.
-
-## The producer will read from the SQS, get the object_key and read the content of S#.
-
-## Consumer demands processor, processor demands producer, producer checks the SQS then reads the S3.
-
-## Buffer demand until consumers are ready to consume more events.
-##	Keep the :min_demand very low in order to only ask for few files at a time
-
