--- sqs_pipeline/deps/gen_stage/hex_metadata.config
@@ -1,21 +1,0 @@
-{<<"app">>,<<"gen_stage">>}.
-{<<"build_tools">>,[<<"mix">>]}.
-{<<"description">>,
- <<"Producer and consumer actors with back-pressure for Elixir">>}.
-{<<"elixir">>,<<"~> 1.5">>}.
-{<<"files">>,
- [<<"lib">>,<<"lib/.DS_Store">>,<<"lib/gen_stage">>,
-  <<"lib/gen_stage/.DS_Store">>,<<"lib/gen_stage/stream.ex">>,
-  <<"lib/gen_stage/streamer.ex">>,<<"lib/gen_stage/dispatcher.ex">>,
-  <<"lib/gen_stage/dispatchers">>,
-  <<"lib/gen_stage/dispatchers/broadcast_dispatcher.ex">>,
-  <<"lib/gen_stage/dispatchers/demand_dispatcher.ex">>,
-  <<"lib/gen_stage/dispatchers/partition_dispatcher.ex">>,
-  <<"lib/gen_stage/utils.ex">>,<<"lib/gen_stage/buffer.ex">>,
-  <<"lib/gen_stage.ex">>,<<"lib/consumer_supervisor.ex">>,
-  <<".formatter.exs">>,<<"mix.exs">>,<<"README.md">>,<<"CHANGELOG.md">>]}.
-{<<"licenses">>,[<<"Apache 2">>]}.
-{<<"links">>,[{<<"GitHub">>,<<"https://github.com/elixir-lang/gen_stage">>}]}.
-{<<"name">>,<<"gen_stage">>}.
-{<<"requirements">>,[]}.
-{<<"version">>,<<"1.0.0">>}.
