--- sqs_pipeline/lib/gtube.ex
@@ -1,18 +1,0 @@
-defmodule Gtube do
-  @moduledoc """
-  Documentation for `Gtube`.
-  """
-
-  @doc """
-  Hello world.
-
-  ## Examples
-
-      iex> Gtube.hello()
-      :world
-
-  """
-  def hello do
-    :world
-  end
-end
