--- sqs_pipeline/test/gtube_test.exs
@@ -1,8 +1,0 @@
-defmodule GtubeTest do
-  use ExUnit.Case
-  doctest Gtube
-
-  test "greets the world" do
-    assert Gtube.hello() == :world
-  end
-end
